Maybe too much, but trying a better infrastructure
This still may only work if it's first connected.  Not sure how the enable environment will handle a thingy = None on setup.  See if this works, and if it works when it's first connected, try throwing the environment stuff in the localThingy class as a do on successful connect.  I think we can still set the notification delegate despite there being nothing. I think it just adds on to our None with information from sensors (which exists).