Somewhat reconnecting
OK, so it's to a place now where it will reconnect for both as an already made and as a off from the start mode.  The problem now is that it seems that the sensor that it reconnects on never gets out of the unavailable state. I'm also not 100% sure the battery sensor works on starting off from nothing.  Something to look into. Also may want to clean up the localThingy init variables.  Not sure if I need all of that.  Look back into dictionaries once it all works.